requirements.txt
fastapi
uvicorn
motor
firebase-admin
PyJWT
python-dotenv
pytest
pytest-asyncio
httpx
pytest-mock
